calculate.js: add mortgage calculations
Adds the ability to simulate taking on a mortgage X months in the future, and simulates a mortgage for every month from today until requirement.

To simulate, an object is created that stores the calculated expected mortgage parameters for the savings saved up to that point in the future.

If the savings are greater than the cost of the house and the taxes, the simulation sets everything to 0.