feat(calculate): stop simulations with enough savings
If a user has enough savings to buy a house without financing, we can stop the simulation.

However, if the savings are too low basically never allow buying a house in the users lifetime, we should stop the simulation.

The simulations now stop getting generated either when the savings are enough to buy the house including taxes, OR when the person is 120 years old, which is safe to assume they will be dead...