refact: use moments date library
Comparing dates in plain javascript is annoying, but the there is a simple library available called moments.js.

This makes difference calculations easier and has better printing abilities for differences.

This also means we can deprecate the months difference function as it was there just to deal with Javascript Date objects