feat: store input values in cookies
Now that we have input validation for the input checkboxes themselves, we can reconsider how the parameters are read in and stored.

We now want to store the validated inputs in cookies and read the stored values on load.
But URL parameters have precedence, so if present the input will be set to that and not the cookie value.

The URL parameters themselves don't really need separate validation so we now just validate after reading either the URL parameter or the cookie value, meaning we can remove the custom URL parameter reading functions.