feat: add house price inflation
The calculations assumed you'd want to buy a house of the same price forever.

But really, 5 years from now you'll want a house at the same "quality" but whose value will have risen with the real estate price inflation.

This adds a house price inflation that gets adjusted monthly.
The inflated price is also communicated to the user.

Now that the house price changes over time, we also consider it part of the total cost calculation.

With a changing house price, the tax also increases, so we need to calculate the tax per simulation period as well.

Relates to #13