refact: move the parameters function to .html
The exportAsURL function was very tied to the structure of the index.html file and the parameters.js file only had one function anyway.

To make it easier to keep the appendParameter and setParameter functions that need to operate on the same list of elements, we moved it to the index.html file.