feat: Add taxes for Portugal
Adds January 2025 tax calculation for continental portugal.

Assumptions are that this is am urban building for primary residence, otherwise, the calculation is different.

Unfortunately, portuguese tax introduces a circular dependency as the mortgage stamp duty applies on the mortgage principal.
As it is a low value we only provide a rough estimate of what that is.

Close #11